refactor Mixins
