+++ mne/channels/channels.py
@@ -19,23 +19,20 @@ from functools import partial
 import numpy as np
 
 from ..defaults import HEAD_SIZE_DEFAULT, _handle_default
-from ..transforms import _frame_to_str
 from ..utils import (verbose, logger, warn,
                      _check_preload, _validate_type, fill_doc, _check_option,
                      _get_stim_channel, _check_fname, _check_dict_keys)
-from ..io.compensator import get_current_comp
 from ..io.constants import FIFF
 from ..io.meas_info import (anonymize_info, Info, MontageMixin, create_info,
                             _rename_comps)
 from ..io.pick import (channel_type, pick_info, pick_types, _picks_by_type,
                        _check_excludes_includes, _contains_ch_type,
                        channel_indices_by_type, pick_channels, _picks_to_idx,
-                       _get_channel_types, get_channel_type_constants,
+                       get_channel_type_constants,
                        _pick_data_channels)
 from ..io.tag import _rename_list
 from ..io.write import DATE_NONE
 from ..io.proj import setup_proj
-from ..io._digitization import _get_data_as_dict_from_dig
 
 
 def _get_meg_system(info):
@@ -189,102 +186,6 @@ def equalize_channels(instances, copy=True, verbose=None):
         logger.info('Channels have been re-ordered.')
 
     return equalized_instances
-
-
-class ContainsMixin(object):
-    """Mixin class for Raw, Evoked, Epochs."""
-
-    def __contains__(self, ch_type):
-        """Check channel type membership.
-
-        Parameters
-        ----------
-        ch_type : str
-            Channel type to check for. Can be e.g. 'meg', 'eeg', 'stim', etc.
-
-        Returns
-        -------
-        in : bool
-            Whether or not the instance contains the given channel type.
-
-        Examples
-        --------
-        Channel type membership can be tested as::
-
-            >>> 'meg' in inst  # doctest: +SKIP
-            True
-            >>> 'seeg' in inst  # doctest: +SKIP
-            False
-
-        """
-        if ch_type == 'meg':
-            has_ch_type = (_contains_ch_type(self.info, 'mag') or
-                           _contains_ch_type(self.info, 'grad'))
-        else:
-            has_ch_type = _contains_ch_type(self.info, ch_type)
-        return has_ch_type
-
-    @property
-    def compensation_grade(self):
-        """The current gradient compensation grade."""
-        return get_current_comp(self.info)
-
-    @fill_doc
-    def get_channel_types(self, picks=None, unique=False, only_data_chs=False):
-        """Get a list of channel type for each channel.
-
-        Parameters
-        ----------
-        %(picks_all)s
-        unique : bool
-            Whether to return only unique channel types. Default is ``False``.
-        only_data_chs : bool
-            Whether to ignore non-data channels. Default is ``False``.
-
-        Returns
-        -------
-        channel_types : list
-            The channel types.
-        """
-        return _get_channel_types(self.info, picks=picks, unique=unique,
-                                  only_data_chs=only_data_chs)
-
-    @fill_doc
-    def get_montage(self):
-        """Get a DigMontage from instance.
-
-        Returns
-        -------
-        %(montage)s
-        """
-        from ..channels.montage import make_dig_montage
-        if self.info['dig'] is None:
-            return None
-        # obtain coord_frame, and landmark coords
-        # (nasion, lpa, rpa, hsp, hpi) from DigPoints
-        montage_bunch = _get_data_as_dict_from_dig(self.info['dig'])
-        coord_frame = _frame_to_str.get(montage_bunch.coord_frame)
-
-        # get the channel names and chs data structure
-        ch_names, chs = self.info['ch_names'], self.info['chs']
-        picks = pick_types(self.info, meg=False, eeg=True, seeg=True,
-                           ecog=True, dbs=True, fnirs=True, exclude=[])
-
-        # channel positions from dig do not match ch_names one to one,
-        # so use loc[:3] instead
-        ch_pos = {ch_names[ii]: chs[ii]['loc'][:3] for ii in picks}
-
-        # create montage
-        montage = make_dig_montage(
-            ch_pos=ch_pos,
-            coord_frame=coord_frame,
-            nasion=montage_bunch.nasion,
-            lpa=montage_bunch.lpa,
-            rpa=montage_bunch.rpa,
-            hsp=montage_bunch.hsp,
-            hpi=montage_bunch.hpi,
-        )
-        return montage
 
 
 channel_type_constants = get_channel_type_constants()
